Move FFT functions to fft.{c,h} files
