+++ src/ckzg.c
@@ -17,4 +17,5 @@
 #include "common.c"
 #include "eip4844.c"
 #include "eip7594.c"
+#include "fft.c"
 #include "setup.c"
