+++ src/eip7594.c
@@ -18,6 +18,7 @@
 
 #include "blst.h"
 #include "common.h"
+#include "fft.h"
 
 #include <inttypes.h>
 #include <stdio.h>
@@ -105,92 +106,6 @@ static bool fr_is_null(const fr_t *p) {
 static uint32_t reverse_bits_limited(uint32_t n, uint32_t value) {
     size_t unused_bit_len = 32 - log2_pow2(n);
     return reverse_bits(value) >> unused_bit_len;
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-// Fast Fourier Transform
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * Fast Fourier Transform.
- *
- * Recursively divide and conquer.
- *
- * @param[out]  out             The results (length `n`)
- * @param[in]   in              The input data (length `n * stride`)
- * @param[in]   stride          The input data stride
- * @param[in]   roots           Roots of unity (length `n * roots_stride`)
- * @param[in]   roots_stride    The stride interval among the roots of unity
- * @param[in]   n               Length of the FFT, must be a power of two
- */
-static void fr_fft_fast(
-    fr_t *out, const fr_t *in, size_t stride, const fr_t *roots, size_t roots_stride, size_t n
-) {
-    size_t half = n / 2;
-    if (half > 0) {
-        fr_t y_times_root;
-        fr_fft_fast(out, in, stride * 2, roots, roots_stride * 2, half);
-        fr_fft_fast(out + half, in + stride, stride * 2, roots, roots_stride * 2, half);
-        for (size_t i = 0; i < half; i++) {
-            blst_fr_mul(&y_times_root, &out[i + half], &roots[i * roots_stride]);
-            blst_fr_sub(&out[i + half], &out[i], &y_times_root);
-            blst_fr_add(&out[i], &out[i], &y_times_root);
-        }
-    } else {
-        *out = *in;
-    }
-}
-
-/**
- * The entry point for forward FFT over field elements.
- *
- * @param[out]  out The results (array of length n)
- * @param[in]   in  The input data (array of length n)
- * @param[in]   n   Length of the arrays
- * @param[in]   s   The trusted setup
- *
- * @remark The array lengths must be a power of two.
- * @remark Use fr_ifft() for inverse transformation.
- */
-static C_KZG_RET fr_fft(fr_t *out, const fr_t *in, size_t n, const KZGSettings *s) {
-    /* Ensure the length is valid */
-    if (n > s->max_width || !is_power_of_two(n)) {
-        return C_KZG_BADARGS;
-    }
-
-    size_t stride = s->max_width / n;
-    fr_fft_fast(out, in, 1, s->expanded_roots_of_unity, stride, n);
-
-    return C_KZG_OK;
-}
-
-/**
- * The entry point for inverse FFT over field elements.
- *
- * @param[out]  out The results (array of length n)
- * @param[in]   in  The input data (array of length n)
- * @param[in]   n   Length of the arrays
- * @param[in]   s   The trusted setup
- *
- * @remark The array lengths must be a power of two.
- * @remark Use fr_fft() for forward transformation.
- */
-static C_KZG_RET fr_ifft(fr_t *out, const fr_t *in, size_t n, const KZGSettings *s) {
-    /* Ensure the length is valid */
-    if (n > s->max_width || !is_power_of_two(n)) {
-        return C_KZG_BADARGS;
-    }
-
-    size_t stride = s->max_width / n;
-    fr_fft_fast(out, in, 1, s->reverse_roots_of_unity, stride, n);
-
-    fr_t inv_len;
-    fr_from_uint64(&inv_len, n);
-    blst_fr_inverse(&inv_len, &inv_len);
-    for (size_t i = 0; i < n; i++) {
-        blst_fr_mul(&out[i], &out[i], &inv_len);
-    }
-    return C_KZG_OK;
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
